+++ R/as_bibentry.R
@@ -15,8 +15,8 @@
 #' object and performs a mapping of the metadata to BibTeX, according to
 #' `vignette("bibtex_cff", "cffr")`.
 #'
-#' The inverse transformation (`bibtex` object to `cff` reference) can be done
-#' with [cff_parse_citation()].
+#' The inverse transformation (`bibentry` object to `cff` reference) can
+#' be done with the corresponding [as_cff()] method.
 #'
 #' @seealso
 #' [utils::bibentry()]